Optimise import time for ast
